# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: antsxdp2.0 admin  接口
  version: 2.0.0
  title: antsxdp 2.0 -admin
basePath: /antsxdp
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /sys/schedule/list:
    post:
      tags:
        - 定时任务
      summary: 定时任务列表
      produces:
        - application/json
      parameters:
        - name: body
          description: map
          in: body
          type: string
          required: true
          default: "{\"page\":\"1\",\"limit\":\"10\",\"beanName\":\"\"}"
      responses:
        '200':
          description: 返回定时任务列表
          schema:
            $ref: '#/definitions/ScheduleJobEntityList'
  /sys/schedule/info/{jobId}:
    get:
      tags:
        - 定时任务
      summary: 获取定时任务信息
      produces:
        - application/json
      parameters:
        - name: jobId
          description: 定时任务ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回定时任务信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              schedule:
                description: 定时任务信息
                $ref: '#/definitions/ScheduleJobEntity'
  /sys/schedule/save:
    post:
      tags:
        - 定时任务
      summary: 添加定时任务
      produces:
        - application/json
      parameters:
        - name: body
          description: 定时任务对象
          in: body
          type: string
          default: "{\"beanName\":\"curlrequesturl\",\"methodName\":\"curlrequesturl\",\"params\":\"http://www.cdn.com\",\"cronExpression\":\"* * * * * ?\",\"status\":0,\"remark\":\"url\"}"
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/update:
    post:
      tags:
        - 定时任务
      summary: 修改定时任务
      produces:
        - application/json
      parameters:
        - name: body
          description: 定时任务对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/ScheduleJobEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/delete:
    post:
      tags:
        - 定时任务
      summary: 删除定时任务
      produces:
        - application/json
      parameters:
        - name: body
          description: ID列表
          in: body
          type: string
          default: "{\"ids\":\"1,2,3\"}"
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/run:
    post:
      tags:
        - 定时任务
      summary: 立即执行任务
      produces:
        - application/json
      parameters:
        - name: body
          description: ""
          default: "{\"ids\":\"1,2,3\"}"
          in: body
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/pause:
    post:
      tags:
        - 定时任务
      summary: 暂停定时任务
      produces:
        - application/json
      parameters:
        - name: body
          description: ""
          default: "{\"ids\":\"1,2,3\"}"
          in: body
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/resume:
    post:
      tags:
        - 定时任务
      summary: 恢复定时任务
      produces:
        - application/json
      parameters:
        - name: body
          description: ""
          default: "{\"ids\":\"1,2,3\"}"
          in: body
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/cdnsys/stream/proxy/list:
    post:
      tags:
        - 端口转发管理
      summary: 分页获取
      produces:
        - application/json
      parameters:
        - name: body
          description: ""
          in: body
          required: true
          default: "{\"page\":1,\"limit\":10,\"user\":\"a\"}"
          type: string
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/cdnsys/stream/proxy/sava:
    post:
      tags:
        - 端口转发管理
      summary: 保存
      produces:
        - application/json
      parameters:
        - name: body
          description: ""
          in: body
          required: true
          default:  "{\"id\":0,\"userId\":61,\"serialNumber\":\"1650954704735002\",\"status\":1,\"confInfo\":\"{\\\"listen\\\":1456,\\\"server\\\":[\\\"172.16.1.5:80 weight=5 max_fails=3 fail_timeout=30s\\\",\\\"2.2.2.2:880\\\"],\\\"server_mode\\\":\\\"hash $remote_addr consistent\\\",\\\"proxy_connect_timeout\\\":\\\"60s\\\",\\\"proxy_timeout\\\":\\\"30s\\\"}\"}"
          type: string
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/cdnsys/stream/proxy/batch/delete:
    post:
      tags:
        - 端口转发管理
      summary: 批量删除
      produces:
        - application/json
      parameters:
        - name: body
          description: ""
          in: body
          required: true
          default: "{\"ids\":\"1,2,3\"}"
          type: string
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/cdnsys/stream/proxy/detail:
    get:
      tags:
        - 端口转发管理
      summary: detail
      produces:
        - application/json
      parameters:
        - name: id
          description: "id"
          in: query
          required: true
          default: 1
          type: integer
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/cdnsys/ip/control/list:
    post:
      tags:
        - IP管理
      summary: ip列表
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          type: string
          description: "control:-1=所有,0=未定义,1=白名单（7层）,2=7层黑名单,3=3层黑名单"
          default: "{\"parentId\":0,\"page\":1,\"limit\":10,\"control\":-1,\"ip\":\"\"}"
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/cdnsys/ip/control/save:
    post:
      tags:
        - IP管理
      summary: ip（分组） 保存与修改
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          type: string
          description:
          default: "{\"id\":0,\"parentId\":0,\"ip\":\"1.1.1.1\",\"control\":1,\"status\":0,\"remark\":\"\",,\"ips\":\"1.1.1.1,1.2.2.2\"}"
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/cdnsys/ip/control/batch/delete:
    post:
      tags:
        - IP管理
      summary: ip删除
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          type: string
          description:
          default: "{\"ids\":\"1,2,3\"}"
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/tb/work/order/category/list:
    get:
      tags:
        - 工单管理
      summary: 查询分类
      produces:
        - application/json
      parameters:
        - name: parentId
          description:  父iD
          in: query
          type: integer
          default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/tb/work/order/wol/list:
    post:
      tags:
        - 工单管理
      summary: 工单列表
      produces:
        - application/json
      parameters:
        - name: body
          description: json
          in: body
          type: string
          default: "{\"urgentLevel\":1,\"status\":0,\"page\":1,\"limit\":20}"
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/tb/work/order/wol/detail:
    get:
      tags:
        - 工单管理
      summary: 工单详情
      produces:
        - application/json
      parameters:
        - name: id
          description: id
          in: query
          type: integer
          default: 1
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/tb/work/order/wol/bat/delete:
    post:
      tags:
        - 工单管理
      summary: 工单批量删除
      produces:
        - application/json
      parameters:
        - name: body
          description: json
          in: body
          type: string
          default: "{\"ids\":\"1,2\"}"
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
      type: object
      properties:
        code:
          description: 状态码 0：成功  非0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
...
  ScheduleJobEntity:
    type: object
    properties:
      jobId:
        description: 任务ID
        type: integer
        format: int64
      beanName:
        description: spring bean名称
        type: string
      methodName:
        description: 方法名
        type: string
      params:
        description: 参数
        type: string
      cronExpression:
        description: cron表达式
        type: string
      status:
        description: 任务状态  0：正常  1：暂停
        type: integer
        format: int32
      remark:
        description: 备注
        type: string
      createTime:
        description: 创建时间
        type: string
        format: date-time
  ScheduleJobEntityList:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/ScheduleJobEntity'
...